fix(change-email): Updated with feedback from peers
