+++ src/Core/Auth/UserFeatures/UserEmail/Interfaces/IChangeEmailCommand.cs
@@ -1,0 +1,8 @@
+﻿using Bit.Core.Entities;
+
+namespace Bit.Core.Auth.UserFeatures.UserEmail.Interfaces;
+
+public interface IChangeEmailCommand
+{
+    Task ChangeEmailAsync(User user, string newEmail, bool logOutUser = true);
+}
